♻️refactor: Widgets are now well organized
⭐ Main Changes:

> Widgets are now organized in folders for better structure according to the type of widget they are. For instance, Forms, Textfields or Buttons are some examples.
